+++ src/Die.h
@@ -70,14 +70,14 @@ namespace NVDSim{
 			Buffer *buffer;
 			Logger *log;
 			bool sending;
-			uint dataCyclesLeft; //cycles per device beat
-			uint deviceBeatsLeft; //device beats per page
-			uint critBeat; //device beat when first cache line will have been sent, used for crit line first
+			uint64_t dataCyclesLeft; //cycles per device beat
+			uint64_t deviceBeatsLeft; //device beats per page
+			uint64_t critBeat; //device beat when first cache line will have been sent, used for crit line first
 			std::queue<ChannelPacket *> returnDataPackets;
 			std::queue<ChannelPacket *> pendingDataPackets;
 			std::vector<Plane> planes;
 			std::vector<ChannelPacket *> currentCommands;
-			uint *controlCyclesLeft;
+			uint64_t *controlCyclesLeft;
 	};
 }
 #endif
