+++ src/GCLogger.cpp
@@ -57,7 +57,7 @@ void GCLogger::update()
 {
     	//update idle energy
 	//since this is already subtracted from the access energies we just do it every time
-	for(uint i = 0; i < (NUM_PACKAGES); i++)
+	for(uint64_t i = 0; i < (NUM_PACKAGES); i++)
 	{
 	  idle_energy[i] += STANDBY_I;
 	}
@@ -203,7 +203,7 @@ void GCLogger::ftlQueueReset()
     gc_queue_length = 0;
 }
 
-void GCLogger::save(uint64_t cycle, uint epoch) 
+void GCLogger::save(uint64_t cycle, uint64_t epoch) 
 {
         // Power stuff
 	// Total power used
@@ -215,7 +215,7 @@ void GCLogger::save(uint64_t cycle, uint epoch)
 	vector<double> ave_erase_power = vector<double>(NUM_PACKAGES, 0.0);
 	vector<double> average_power = vector<double>(NUM_PACKAGES, 0.0);
 
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{
 	    if(cycle != 0)
 	    {
@@ -301,9 +301,9 @@ void GCLogger::save(uint64_t cycle, uint epoch)
 	savefile<<"========================\n";
 	savefile<<"Maximum Length of Ftl Queue: " <<max_ftl_queue_length<<"\n";
 	savefile<<"Maximum Length of GC Queue: " <<max_gc_queue_length<<"\n";
-	for(uint i = 0; i < max_ctrl_queue_length.size(); i++)
+	for(uint64_t i = 0; i < max_ctrl_queue_length.size(); i++)
 	{
-	    for(uint j = 0; j < max_ctrl_queue_length[i].size(); j++)
+	    for(uint64_t j = 0; j < max_ctrl_queue_length[i].size(); j++)
 	    {
 		savefile<<"Maximum Length of Controller Queue for Package " << i << ", Die " << j << ": "<<max_ctrl_queue_length[i][j]<<"\n";
 	    }
@@ -323,7 +323,7 @@ void GCLogger::save(uint64_t cycle, uint epoch)
 	savefile<<"\nPower Data: \n";
 	savefile<<"========================\n";
 
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{
 	    savefile<<"Package: "<<i<<"\n";
 	    savefile<<"Accumulated Idle Energy: "<<(idle_energy[i] * VCC * (CYCLE_TIME * 0.000000001))<<" mJ\n";
@@ -362,7 +362,7 @@ void GCLogger::print(uint64_t cycle) {
 	vector<double> ave_erase_power = vector<double>(NUM_PACKAGES, 0.0);
 	vector<double> average_power = vector<double>(NUM_PACKAGES, 0.0);
 
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{
 	  total_energy[i] = (idle_energy[i] + access_energy[i] + erase_energy[i]) * VCC;
 	  ave_idle_power[i] = (idle_energy[i] * VCC) / cycle;
@@ -378,7 +378,7 @@ void GCLogger::print(uint64_t cycle) {
 	cout<<"\nPower Data: \n";
 	cout<<"========================\n";
 
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{
 	    cout<<"Package: "<<i<<"\n";
 	    cout<<"Accumulated Idle Energy: "<<(idle_energy[i] * VCC * (CYCLE_TIME * 0.000000001))<<"mJ\n";
@@ -398,7 +398,7 @@ void GCLogger::print(uint64_t cycle) {
 vector<vector<double> > GCLogger::getEnergyData(void)
 {
     vector<vector<double> > temp = vector<vector<double> >(3, vector<double>(NUM_PACKAGES, 0.0));
-    for(uint i = 0; i < NUM_PACKAGES; i++)
+    for(uint64_t i = 0; i < NUM_PACKAGES; i++)
     {
 	temp[0][i] = idle_energy[i];
 	temp[1][i] = access_energy[i];
@@ -407,7 +407,7 @@ vector<vector<double> > GCLogger::getEnergyData(void)
     return temp;
 }
 
-void GCLogger::save_epoch(uint64_t cycle, uint epoch)
+void GCLogger::save_epoch(uint64_t cycle, uint64_t epoch)
 {    
     EpochEntry this_epoch;
     this_epoch.cycle = cycle;
@@ -441,15 +441,15 @@ void GCLogger::save_epoch(uint64_t cycle, uint epoch)
 
     this_epoch.writes_per_address = writes_per_address;
     
-    for(uint i = 0; i < ctrl_queue_length.size(); i++)
+    for(uint64_t i = 0; i < ctrl_queue_length.size(); i++)
     {
-	for(uint j = 0; j < ctrl_queue_length[i].size(); j++)
+	for(uint64_t j = 0; j < ctrl_queue_length[i].size(); j++)
 	{
 	    this_epoch.ctrl_queue_length[i][j] = ctrl_queue_length[i][j];
 	}
     }
 
-    for(uint i = 0; i < NUM_PACKAGES; i++)
+    for(uint64_t i = 0; i < NUM_PACKAGES; i++)
     {	
 	this_epoch.idle_energy[i] = idle_energy[i]; 
 	this_epoch.access_energy[i] = access_energy[i]; 
@@ -486,7 +486,7 @@ void GCLogger::save_epoch(uint64_t cycle, uint epoch)
 	this_epoch.average_gcwrite_latency -= last_epoch.average_gcwrite_latency;
 	this_epoch.average_queue_latency -= last_epoch.average_queue_latency;
     
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{	
 	    this_epoch.idle_energy[i] -= last_epoch.idle_energy[i]; 
 	    this_epoch.access_energy[i] -= last_epoch.access_energy[i]; 
@@ -541,7 +541,7 @@ void GCLogger::write_epoch(EpochEntry *e)
 	vector<double> ave_erase_power = vector<double>(NUM_PACKAGES, 0.0);
 	vector<double> average_power = vector<double>(NUM_PACKAGES, 0.0);
 
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{
 	    if(e->cycle != 0)
 	    {
@@ -601,9 +601,9 @@ void GCLogger::write_epoch(EpochEntry *e)
 	savefile<<"========================\n";
 	savefile<<"Length of Ftl Queue: " <<e->ftl_queue_length<<"\n";
 	savefile<<"Length of GC Queue: " <<e->gc_queue_length<<"\n";
-	for(uint i = 0; i < e->ctrl_queue_length.size(); i++)
+	for(uint64_t i = 0; i < e->ctrl_queue_length.size(); i++)
 	{
-	    for(uint j = 0; j < e->ctrl_queue_length[i].size(); j++)
+	    for(uint64_t j = 0; j < e->ctrl_queue_length[i].size(); j++)
 	    {
 		savefile<<"Length of Controller Queue for Package " << i << ", Die " << j << ": "<<e->ctrl_queue_length[i][j]<<"\n";
 	    }
@@ -623,7 +623,7 @@ void GCLogger::write_epoch(EpochEntry *e)
 	savefile<<"\nPower Data: \n";
 	savefile<<"========================\n";
 
-	for(uint i = 0; i < NUM_PACKAGES; i++)
+	for(uint64_t i = 0; i < NUM_PACKAGES; i++)
 	{
 	    savefile<<"Package: "<<i<<"\n";
 	    savefile<<"Accumulated Idle Energy: "<<(e->idle_energy[i] * VCC * (CYCLE_TIME * 0.000000001))<<" mJ\n";
